\name{print.pnardl}
\alias{print.pnardl}
\title{Print Method for pnardl Objects}
\usage{
\method{print}{pnardl}(x, digits = 4L, ...)
}
\arguments{
\item{x}{An object of class \code{"pnardl"}.}

\item{digits}{Integer. Number of significant digits to display.}

\item{...}{Further arguments passed to or from other methods (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Method for pnardl Objects
}
